EventDisplay/requirements.txt

cycler==0.10.0
kiwisolver==1.1.0
matplotlib>=3.1.3
numpy>=1.18.1
opencv-python>=4.2.0.32
pandas==2.0.3
Pillow>=7.0.0
pyparsing==2.4.6
python-dateutil>=2.8.1
pytz>=2019.3
scipy>=1.4.1
six>=1.14.0
